+++ test/unittest.hh
@@ -18,26 +18,26 @@
 */
 
 #ifndef TEST_UNITTEST_HH
-#  define TEST_UNITTEST_HH
+#define TEST_UNITTEST_HH
 
-#  include <cstdio>
-#  include <cstdlib>
-#  include <iostream>
-#  include "com/centreon/clib.hh"
-#  include "com/centreon/engine/broker/compatibility.hh"
-#  include "com/centreon/engine/broker/loader.hh"
-#  include "com/centreon/engine/checks/checker.hh"
-#  include "com/centreon/engine/commands/set.hh"
-#  include "com/centreon/engine/configuration/applier/state.hh"
-#  include "com/centreon/engine/configuration/state.hh"
-#  include "com/centreon/engine/events/loop.hh"
-#  include "com/centreon/engine/globals.hh"
-#  include "com/centreon/engine/logging/logger.hh"
-#  include "com/centreon/engine/namespace.hh"
-#  include "com/centreon/engine/timezone_manager.hh"
-#  include "com/centreon/logging/backend.hh"
-#  include "com/centreon/logging/engine.hh"
-#  include "com/centreon/logging/file.hh"
+#include <cstdio>
+#include <cstdlib>
+#include <iostream>
+#include "com/centreon/clib.hh"
+#include "com/centreon/engine/broker/compatibility.hh"
+#include "com/centreon/engine/broker/loader.hh"
+#include "com/centreon/engine/checks/checker.hh"
+#include "com/centreon/engine/commands/set.hh"
+#include "com/centreon/engine/configuration/applier/state.hh"
+#include "com/centreon/engine/configuration/state.hh"
+#include "com/centreon/engine/events/loop.hh"
+#include "com/centreon/engine/globals.hh"
+#include "com/centreon/engine/logging/logger.hh"
+#include "com/centreon/engine/namespace.hh"
+#include "com/centreon/engine/timezone_manager.hh"
+#include "com/centreon/logging/backend.hh"
+#include "com/centreon/logging/engine.hh"
+#include "com/centreon/logging/file.hh"
 
 CCE_BEGIN()
 
@@ -47,8 +47,8 @@ CCE_BEGIN()
  *  engine needs to make unit test and run
  *  unit test.
  */
-class     unittest {
-public:
+class unittest {
+ public:
   /**
    *  Constructor.
    *
@@ -58,23 +58,20 @@ public:
    *
    *  @return Return value of func.
    */
-             unittest(
-               int argc,
-               char** argv,
-               int (* func)(int, char**))
-    : _argc(argc), _argv(argv), _func(func), _log(stdout) {}
+  unittest(int argc, char** argv, int (*func)(int, char**))
+      : _argc(argc), _argv(argv), _func(func), _log(stdout) {}
 
   /**
    *  Destructor.
    */
-             ~unittest() throw () {}
+  ~unittest() throw() {}
 
   /**
    *  Entry point.
    *
    *  @return Return value of unit test routine.
    */
-  int        run() {
+  int run() {
     int ret(EXIT_FAILURE);
     if (!_init())
       return (ret);
@@ -92,41 +89,39 @@ public:
     return (ret);
   }
 
-private:
-  class      nothing
-    : public com::centreon::logging::backend {
-  public:
-             nothing(FILE* file = NULL) { (void)file; }
-             nothing(nothing const& right)
-      : com::centreon::logging::backend(right) { (void)right; }
-             ~nothing() throw () {}
+ private:
+  class nothing : public com::centreon::logging::backend {
+   public:
+    nothing(FILE* file = NULL) { (void)file; }
+    nothing(nothing const& right) : com::centreon::logging::backend(right) {
+      (void)right;
+    }
+    ~nothing() throw() {}
     nothing& operator=(nothing const& right) {
       (void)right;
       return (*this);
     }
-    void     close() throw () {}
-    void     log(
-               unsigned long long types,
-               unsigned int verbose,
-               char const* msg,
-               unsigned int size) throw () {
+    void close() throw() {}
+    void log(uint64_t types,
+             uint32_t verbose,
+             char const* msg,
+             uint32_t size) throw() {
       (void)types;
       (void)verbose;
       (void)msg;
       (void)size;
     }
-    void     open() {}
-    void     reopen() {}
+    void open() {}
+    void reopen() {}
   };
 
-  bool       _init() {
+  bool _init() {
     try {
       com::centreon::clib::load();
-      com::centreon::logging::engine::instance()
-        .add(
-           &_log,
-           com::centreon::engine::logging::log_all,
-           com::centreon::engine::logging::most);
+      com::centreon::logging::engine::instance().add(
+          &_log,
+          com::centreon::engine::logging::log_all,
+          com::centreon::engine::logging::most);
       config = new configuration::state;
       timezone_manager::load();
       commands::set::load();
@@ -137,19 +132,18 @@ private:
       broker::compatibility::load();
     }
     catch (std::exception const& e) {
-      std::cerr << "unit test init failed: "
-                << e.what() << std::endl;
+      std::cerr << "unit test init failed: " << e.what() << std::endl;
       return (false);
     }
     catch (...) {
       std::cerr << "unit test init failed: "
-        "unknown exception" << std::endl;
+                   "unknown exception" << std::endl;
       return (false);
     }
     return (true);
   }
 
-  bool       _deinit() {
+  bool _deinit() {
     try {
       broker::compatibility::unload();
       broker::loader::unload();
@@ -162,29 +156,27 @@ private:
       com::centreon::clib::unload();
     }
     catch (std::exception const& e) {
-      std::cerr << "unit test deinit failed: "
-                << e.what() << std::endl;
+      std::cerr << "unit test deinit failed: " << e.what() << std::endl;
       return (false);
     }
     catch (...) {
       std::cerr << "unit test deinit failed: "
-        "unknown exception" << std::endl;
+                   "unknown exception" << std::endl;
       return (false);
     }
     return (true);
   }
 
-  int     _argc;
-  char**  _argv;
-  int     (*_func)(int, char**);
-#  ifdef NDEBUG
+  int _argc;
+  char** _argv;
+  int (*_func)(int, char**);
+#ifdef NDEBUG
   nothing _log;
-#  else
-  com::centreon::logging::file
-          _log;
-#  endif // !NDEBUG
+#else
+  com::centreon::logging::file _log;
+#endif  // !NDEBUG
 };
 
 CCE_END()
 
-#endif // !TEST_UNITTEST_HH
+#endif  // !TEST_UNITTEST_HH
